create helper function for drag and drop
